Games

+++ Text_RPG/MainMenu.h
@@ -3,6 +3,8 @@
 #include <string>
 #include <sstream>
 #include <fstream>
+#include <conio.h>
+#include <Windows.h>
 
 
 using namespace std;
@@ -11,25 +13,23 @@ class MainMenu
 {
 private:
 	
-	unsigned short Menu_Selection;
-
+	unsigned int Menu_Selection;
 	string Menu_Text;
-	
-	
-
 
 public:
 
+	
 	void Init_Variable();
-
-
+	
 	MainMenu();
 
 	virtual ~MainMenu();
 
+	
+
 	void Menu();
 
-
+	
 	virtual void Update();
